example/android/requirements.txt
Appium-Python-Client
selenium==3.141.0
pytest
pytest-xdist
uiautomator
uiautomator2
